code refactor and fix output pathes
