+++ Hnet/hnet_loss_v2.py
@@ -2,7 +2,8 @@ import torch
 import numpy as np
 from torch.nn.modules.loss import _Loss
 
-from hnet_utils import hnet_single_frame_loss
+from hnet_utils import hnet_transformation
+
 
 # Use GPU if available, else use CPU
 device = torch.device(
@@ -40,8 +41,7 @@ class HetLoss(_Loss):
     def forward(self, input_pts, transformation_coefficient):
         return self._hnet_loss(input_pts, transformation_coefficient)
 
-    @staticmethod
-    def _hnet_loss(input_pts, transformation_coefficient):
+    def _hnet_loss(self, input_pts, transformation_coefficient):
         # assert not torch.isnan(transformation_coefficient).any(), "transformation_coefficient is nan"
         # todo: handle case where transformation_coefficient is nan
         if torch.isnan(transformation_coefficient).any():
@@ -53,9 +53,26 @@ class HetLoss(_Loss):
         for i in range(batch_size):
             frame_input_pts = input_pts[i]
             frame_transformation_coefficient = transformation_coefficient[i]
-            frame_loss = hnet_single_frame_loss(frame_input_pts, frame_transformation_coefficient)
+            frame_loss = self.hnet_single_frame_loss(frame_input_pts, frame_transformation_coefficient)
             single_frame_losses.append(frame_loss)
 
         loss = torch.mean(torch.stack(single_frame_losses))
 
         return loss
+
+    @staticmethod
+    def hnet_single_frame_loss(input_pts, transformation_coefficient, poly_fit_order: int = 3):
+        """
+        :param input_pts: the points of the lane of a single image, shape: [k, 3] (k is the number of points)
+        :param transformation_coefficient: the 6 params from the HNet, shape: [1, 6]
+        :param poly_fit_order: the order of the polynomial
+        :return single_frame_loss: the loss of the single frame
+        """
+
+        valid_pts_reshaped, _, preds_transformation_back, _ = hnet_transformation(input_pts,
+                                                                                  transformation_coefficient,
+                                                                                  poly_fit_order)
+        # compute loss between back-transformed polynomial fit and gt_pts
+        single_frame_loss = torch.mean(torch.pow(valid_pts_reshaped[0, :] - preds_transformation_back[0, :], 2))
+
+        return single_frame_loss
